Re-home admin user utility actions
Amp-Thread-ID: https://ampcode.com/threads/T-01a05ac9-27dc-712d-ae46-504e7764121a
