+++ app/controllers/api/admin/v1/application_controller.rb
@@ -2,7 +2,15 @@ module Api
   module Admin
     module V1
       class ApplicationController < Api::Admin::ApplicationController
-        # todo, add shit here or something
+        private
+
+        def find_user_by_id
+          user_id = params[:id] || params[:user_id]
+          return render_error("who?") if user_id.blank?
+          User.find(user_id)
+        rescue ActiveRecord::RecordNotFound
+          render_not_found_json("user not found")
+        end
       end
     end
   end
